parser-json-cov: separate module for CovTreeDecoder
