+++ src/CMakeLists.txt
@@ -76,6 +76,7 @@ add_library(cs STATIC
     parser-cov.cc
     parser-gcc.cc
     parser-json.cc
+    parser-json-cov.cc
     parser-json-simple.cc
     parser-xml.cc
     shared-string.cc
